requirements.txt
fastapi
uvicorn
python-multipart
jinja2
gunicorn
python-dotenv
spotipy
gspread
google-auth-oauthlib
psutil
